requirements.txt
alembic==1.4.3
aniso8601==8.0.0
autopep8==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.22.3
mysqlclient==2.0.1
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
Werkzeug==1.0.1
